// frontend/src/app/core/api/model/vaultEntryRequest.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request object for creating or updating a vault entry
 */
export interface VaultEntryRequest { 
    /**
     * The title or name of the entry
     */
    title?: string;
    /**
     * The username or email for the account
     */
    username?: string;
    /**
     * The password for the account
     */
    password?: string;
    /**
     * The URL of the website
     */
    websiteUrl?: string;
    /**
     * Optional notes regarding the account
     */
    notes?: string;
    /**
     * Optional ID of a Category to assign this entry to
     */
    categoryId?: number;
    /**
     * Optional ID of a Folder to place this entry inside
     */
    folderId?: number;
    /**
     * Boolean flag to mark entry as a favorite
     */
    isFavorite?: boolean;
    /**
     * Boolean flag indicating if the entry is highly sensitive (requires re-prompt of master password)
     */
    isHighlySensitive?: boolean;
}