// frontend/src/app/core/api/model/vaultEntryDetailResponse.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface VaultEntryDetailResponse { 
    id?: number;
    title?: string;
    username?: string;
    password?: string;
    websiteUrl?: string;
    notes?: string;
    categoryId?: number;
    categoryName?: string;
    folderId?: number;
    folderName?: string;
    isFavorite?: boolean;
    isHighlySensitive?: boolean;
    requiresSensitiveAuth?: boolean;
    createdAt?: string;
    updatedAt?: string;
    strengthScore?: number;
    strengthLabel?: string;
}