// frontend/src/app/core/api/model/categoryDTO.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CategoryDTO { 
    id?: number;
    name?: string;
    icon?: string;
    isDefault?: boolean;
    createdAt?: string;
    entryCount?: number;
}